Improved architecture
