+++ components/FooterSection/ContactForm/Input/styles.ts
@@ -1,0 +1,33 @@
+import styled from "styled-components";
+import { Field } from "formik";
+
+export const FormFieldWrapper = styled.div`
+  width: 100%;
+  position: relative;
+`;
+
+interface FormFieldProps {
+  $isError: boolean;
+}
+
+export const FormField = styled(Field)<FormFieldProps>`
+  width: 100%;
+  font-size: 16px;
+  padding-bottom: 23px;
+  border-bottom: 1px solid
+    ${(props: FormFieldProps) =>
+      props.$isError ? "var(--color-red)" : "var(--color-black100)"};
+
+  ::placeholder {
+    color: ${(props: FormFieldProps) =>
+      props.$isError ? "var(--color-red)" : "var(--color-black70)"};
+  }
+`;
+
+export const StyledWarning = styled.i`
+  color: var(--color-red);
+  position: absolute;
+  right: 0;
+  top: 1px;
+  font-size: 20px;
+`;
