+++ data/experiences/index.ts
@@ -1,0 +1,19 @@
+import Experience from "@/types/Experience";
+
+export default [
+  {
+    title: "Full-Stack Developer",
+    company: "Spring — Full Service Studio",
+    period: "Mar 2022 — Jul 2022",
+  },
+  {
+    title: "iOS Engineer",
+    company: "Crowdhaus — Property Discovery",
+    period: "Jan 2021 — Mar 2022",
+  },
+  {
+    title: "Founder",
+    company: "Card — Share Social Profiles",
+    period: "Jun 2020  — Nov 2022",
+  },
+] as Experience[];
