+++ data/projects/index.ts
@@ -1,0 +1,40 @@
+import Project from "@/types/Project";
+
+import JoePeshallImage from "@/public/portfolio/joepeshall.png";
+import SpringImage from "@/public/portfolio/byspring.png";
+import DriftImage from "@/public/portfolio/drift.png";
+import TabagerImage from "@/public/portfolio/tabager.png";
+import ZonloImage from "@/public/portfolio/zonlo.png";
+
+export default [
+  {
+    title: "Joe’s Portfolio Website",
+    subtitle: "Web Development",
+    image: JoePeshallImage,
+    url: "https://www.joepeshall.co.uk",
+  },
+  {
+    title: "Spring — Full Service Studio",
+    subtitle: "Web Development",
+    image: SpringImage,
+    url: "https://www.byspring.co",
+  },
+  {
+    title: "Drift — Focus & Meditation",
+    subtitle: "Mobile Development",
+    image: DriftImage,
+    url: "https://apps.apple.com/us/app/drift-focus-meditation/id1620172749",
+  },
+  {
+    title: "Tabager — Manage Web Tabs",
+    subtitle: "Product Design, Web Development",
+    image: TabagerImage,
+    url: "https://www.tabager.com/register",
+  },
+  {
+    title: "Zonlo — Internships Across UK",
+    subtitle: "Product Design, Web Development",
+    image: ZonloImage,
+    url: "https://zonlo.vercel.app",
+  },
+] as Project[];
