+++ data/socialMediaLinks/index.ts
@@ -1,0 +1,29 @@
+import SocialLink from "@/types/SocialLink";
+
+export default [
+  {
+    iconClass: "ri-linkedin-box-line",
+    url: "https://www.linkedin.com/in/mitk8/",
+    color: "blue",
+  },
+  {
+    iconClass: "ri-github-line",
+    url: "https://github.com/andrey-mitko",
+    color: "pink",
+  },
+  {
+    iconClass: "ri-product-hunt-line",
+    url: "https://www.producthunt.com/@andrey_mitko",
+    color: "green",
+  },
+  {
+    iconClass: "ri-twitter-line",
+    url: "https://twitter.com/andrey_mitko",
+    color: "orange",
+  },
+  {
+    iconClass: "ri-medium-line",
+    url: "https://medium.com/@mitk8",
+    color: "purple",
+  },
+] as SocialLink[];
