+++ next.config.js
@@ -3,11 +3,6 @@ const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
   optimizeFonts: false,
-
-  compiler: {
-    // see https://styled-components.com/docs/tooling#babel-plugin for more info on the options.
-    styledComponents: true,
-  },
 };
 
 module.exports = nextConfig;
