+++ types/Experience/index.ts
@@ -1,0 +1,5 @@
+export default interface Experience {
+  title: string;
+  company: string;
+  period: string;
+}
