+++ types/NavItem/index.ts
@@ -1,0 +1,4 @@
+export default interface NavItem {
+  title: string;
+  url: string;
+}
