+++ types/Project/index.ts
@@ -1,0 +1,8 @@
+import { StaticImageData } from "next/image";
+
+export default interface Project {
+  title: string;
+  subtitle: string;
+  image: StaticImageData;
+  url: string;
+}
