+++ types/SocialLink/index.ts
@@ -1,0 +1,5 @@
+export default interface SocialLink {
+  iconClass: string;
+  url: string;
+  color: string;
+}
